requirements.txt
streamlit
matplotlib
tensorflow
tensorflow.keras
numpy
Pillow
ipython